feat(core): Use a container to compile protos
Compiling the protos requires a number of tools to be installed on
the local machine (protoc, protoc-gen-go, protoc-gen-doc). While
it's not too difficult to set this up once, differences in version
of these tools between different contributors can lead to large diffs
in generated output, and also makes is difficult to write tooling that
ensures that the generated output is in sync.

This commit creates a Dockerfile containing all the required tools at
pinned versions. Now all that is needed to compile the protos reproducibly
is to have docker installed.

It also adds a Makefile to make it easy to regenerate the protos. Running
'make proto' will update all generated code based on the current status of
the proto files. Running 'make checkproto' will check if the generated code
and protos are currently in sync, but will not update any of the generated
code.  (The idea is that 'make checkproto' can be run as part of pre-merge
CI to ensure that these always stay in sync.)