chore(core): Update protoc-gen-go to V2
Go has just released an updated protobuf API implementation; there
are a number of breaking changes that make it incompatible with the
prior version (and the import path has changed from
github.com/golang/protobuf to google.golang.org/protobuf to indicate
this).

Given that we are just starting to implement our protos, and none
of the breaking changes affect us, it probably makes sense to
upgrade now (before we have any code that does rely on behavior
of the old version).

This commit updates us to use the new generator by:
(1) Updating the protoc-gen-go docker container to pull from the
new repository.
(2) Regenerating the generated code.
(3) Updating the go modules to account for new imports in this
new generated code.