fix(cmd): Executable should be in package main
When re-organizing the directories, we renamed the package of
the main binary from main to kleat. While all importable packages
need to be in a directory with the same name of the package, the
entrypoint of the program must be a function main in a package
main.

the executable anymore as go programs mus