refactor(core): Use sigs.k8s.io/yaml instead of ghodss/yaml
There has not been a release of ghodss/yaml in more than 3 years;
it looks like kubernetes has started maintaining a fork of it with
more recent releases, so let's use that.

The motivation for this is that I want to use UnmarshalStrict in
tests so they catch additional fields that shouldn't be there, but
that is not present in the latest release of ghodss/yaml (though is
on the master branch there).