feat(echo): Add functionality to generate echo config
This commit adds an empty echo config, and adds the functionality
for kleat to generate an echo config, and a basic test on that
config.

The code to generate the config (as well as the test) is somewhat
duplicative of the code for clouddriver. For now I'm deferring
cleaning that up into better reusable components, but will do
that before we add a third microservice.