feat(echo): Add notification config (#29)
* refactor(core): Use sigs.k8s.io/yaml instead of ghodss/yaml

There has not been a release of ghodss/yaml in more than 3 years;
it looks like kubernetes has started maintaining a fork of it with
more recent releases, so let's use that.

The motivation for this is that I want to use UnmarshalStrict in
tests so they catch additional fields that shouldn't be there, but
that is not present in the latest release of ghodss/yaml (though is
on the master branch there).

* feat(echo): Add functionality to generate echo config

This commit adds an empty echo config, and adds the functionality
for kleat to generate an echo config, and a basic test on that
config.

The code to generate the config (as well as the test) is somewhat
duplicative of the code for clouddriver. For now I'm deferring
cleaning that up into better reusable components, but will do
that before we add a third microservice.

* feat(echo): Add notification config

This commit adds support for configuring notifications via the
halconfig, and having those be written to the echo config.

As artifacts are just passed-through to echo, and the clouddriver
work already strongly-typed this, this commit also passes artifacts
through to echo.

* fix(echo): Address PR review comments

I guess I'll have to regenerate the protos manually and push another commit after this one.

Co-Authored-By: Maggie Neterval <mneterval@google.com>

Co-authored-by: Maggie Neterval <mneterval@google.com>