feat(echo): Add GCB account
Echo requires GCB accounts to be in its config file, as it listens
to their configured Pub/Sub topics to respond to changes in build
status.

This commit adds top-level CI field to the hal config, and implements
support for GCB accounts. Support for other CI accounts will be done
when we implement the rest of the igor config.