refactor(core): Standardize on using pointers
For a while we weren't sure whether to pass by value or by
reference. Given that we're passing around structs whose nested
fields are all themselves passed around by reference, it seems
logical to do the same at the top level (as it's confusing that the
top-level thing is sometimes passed around by value).

This also resolves a bunch of 'go vet' warnings because the generated
structs contain a mutex somewhere, which really should not be copied,
and generates a warning when we pass these around by value. If I didn't
think that it probably makes sense to pass by value anyway I'd probably
spend more time seeing if that warning shoudl be suppressed, but given
that I think it makes sense anyway to pass by value, let's do that.