feat(echo): Add GCB account and stats (#32)
* feat(echo): Add GCB account

Echo requires GCB accounts to be in its config file, as it listens
to their configured Pub/Sub topics to respond to changes in build
status.

This commit adds top-level CI field to the hal config, and implements
support for GCB accounts. Support for other CI accounts will be done
when we implement the rest of the igor config.

* feat(stats): Add stats to echo config

For now this only passes through whether stats are enabled or not;
we'll need to think more about how to handle generating a unique id
for the instance, and passing through the Spinnaker version.

* fix(gcb): Add comments to GCB protos

Co-authored-by: Maggie Neterval <mneterval@google.com>

Co-authored-by: Maggie Neterval <mneterval@google.com>