chore(kubernetes): Remove version flag from Kubernetes provider
The V1 provider will be removed from the codebase by the time that
kleat is ready for use.

I didn't renumber all the protos to fill the missing number to
reduce the size of the diff (and because at some point we'll
have to stop doing this).  Maybe we can do one last cleanup
before some kind of 'alpha' release.