feat(proto): Use protojson to unmarshal protos (#40)
* feat(proto): Use protojson to unmarshal protos

Instead of relying on the struct tags that are added by protoc-gen-go,
use protojson to marshal/unmarshal objects that implement the
proto.Message interface. This has the advantage of correctly
marshaling/unmarshaling enums, as well as getting us compatibility
with the standard JSON format for proto messages.

As there is no standard YAML format for proto messages (and YAML is
a superset of JSON), there is no protoyaml package. Make a thin
wrapper that converts YAML to JSON before using protojson (and the
reverse for marshaling). This is effectively what our yaml.Unmarshal
was doing before under the hood (to use the JSON struct tags) so
the JSON <-> YAML conversion is unaffected, it's jsut hat the
JSON <-> Object conversion now uses protojson instead of go's
built-in json.

* feat(proto): Add enum types

Now that we can marshal and unmarshal enum types, replace types that
were made strings uniquely to work around the unmarshaling issues.

* chore(kubernetes): Remove version flag from Kubernetes provider

The V1 provider will be removed from the codebase by the time that
kleat is ready for use.

I didn't renumber all the protos to fill the missing number to
reduce the size of the diff (and because at some point we'll
have to stop doing this).  Maybe we can do one last cleanup
before some kind of 'alpha' release.