fix(proto): Don't delete all generated code if protoc fails
We currently delete all the generated code before running protoc
so that we can start with an empty directory (avoiding cases where
we miss deleting an old generated file).

This means that if protoc fails, we delete all the generated code,
which can be annoying. Let's instead generate the code into the
staging directory and only move it into the final directory if
protoc exits successfully.