refactor(proto): Misc cleanup on protos
Add a few missing comments and alphabetize imports.

In protos that only contain other non-primitive types, I didn't
add a comment on each field. (Ex: Given that the Clouddriver proto
has a KubernetesProvider field that is itself commented, it seemed
redundant to also add a comment inside the clouddriver proto.)