refactor(test): Clean up variable names and error logging
In a few places I think we may have gone overboard with short
variable names as I'm often scrolling up to see what a single-letter
variable name means. Let's give these names that are a bit more
descriptive, but still short.

Also, the tests all call t.Errorf when something goes wrong. That
causes execution to continue, which is good if you are checking an
assertion and still want to check later ones, but bad if there's
something fundamentally wrong with the test that makes later
assertions useless. For example, if we can't read in the echo.yml
we should just fail immediately rather than continue and log a failure
that the empty string didn't match what was expected. This changes
any failure that prevents getting anything meaningful from the test
to instead call Fail.  (And also by using .Fail instead of .Failf
we can pass the error directly instead of converting it to a string.)

It's somewhat moot that the last assertion in each test I've left
as .Errorf, because the test is over anyway, but it matches the
principle above---that the test could continue (ie, checking other
conditions) if the assertion failed.