feat(security): Add security to hal config
This commit adds the security block to the hal config, as well
as the configuration nested under it. This includes SSL for gate/deck,
as well as authentication and authorization.

The one exception is that I left a TODO for OAuth as Halyard looks
to be doing some complex logic there and I want to look at separately.