fix(gate): Update SAML proto
Halyard did some mappings and transformations that we won't be
replicating in kleat. Notably, Halyard parsed the metadata string
you passed on the command line to store it in either metadataLocal
or metadataRemote; but then when generating the gate config it would
just add whichever one was not null to metadataUrl. We'll only
support metadataUrl in kleat and users will need to (once) manually
update their key from metadataLocal/metadataRemote to metadataUrl.

Halyard accepted a redirect URI for SAML and parsed out the port,
host, and protocol. We'll require that users supply these separately,
which is how Gate expects them.