feat(gate): Add initial gate config
This commit adds the ability to configure Gate using kleat. I still
need to add tests for gate, as well as do some more extensive comparison
with what Halyard outputs to make sure everything is correct here, but
this defines all of the types and allows a gate config to be generated.