fix(proto): Wrap some proto values (#58)
This commit wraps some proto fields in cases where the service
needs to make a meaningful distinction between the default
value for the type and no value for the type (ie null).

In particular, these are:
* Booleans that currently (or might reasonably in the future)
default to true in the service (and where explicitly passing
false would be different from not passing a value)
* Strings where empty string is a reasonable value but we currently
(or might reasonably in the future) default nulls to something that's
not the empty string.

It also removes a parameter from the Travis config that was being
ignored by the service and so should not be supported in kleat.