fix(proto): Fix some of the security protos
Halyard is writing the edited enums in upper case, so we
should make our enum the same. (I was confused because the
toString was writing it in lower-case, but I guess Halyard is
not using toString as part of serialization, which I guess
probably makes sense when you think about it enough.)

I also accidentally named a property spring instead of security.
Let's also rename the message SpringSecurity and un-nest it as
it represents the top-level security.* config settings for
Spring (to which I'll be shortly adding another field).