feat(proto): Add basic authentication
Halyard currently looks on the service settings file for a flag to
determine whether to enable basic authentication; this is different
from all the other authentication methods, which are configured
together in the same block of the hal config.

As I don't anticipate wanting to expose the service settings through
kleat (as these are in general related to how a service is deployed
rather than how it behaves) let's add a new field to the halconfig
so users can still configure it if they want.

I suspect this is not very common in any real production set-up
(as it would require sharing a username and password with everyone
using Spinnaker) but as it was easy to just add the two fields, it
seemed better than not supporting it anymore. Users *will* need to
move any basic auth config from the service settings file to the
hal config.