feat(gate): Write out the gate config to file
Also, update the logic for generating the gate config to
avoid generating config blocks that are not relevant. In general,
I think we should adopt the pattern (and sort of already have) where
if we're not just looking something up in the hal config and are
actually creating a new struct inline, we should first check to
see if we need to create the wrapper struct to avoid writing out
a bunch of empty structs.

Remove the TODO as upon further investigation we don't need to allow
these to be settable in kleat and can rely on the default port and
binding.