test(gate): Add test for gate config generation
This commit adds the end-to-end test that the gate config is correctly
generated when everything is specified. (Of course this isn't really
realisitic as you likely won't have every type of auth configured but
given that kleat just passes things through it's still a good test.)

Next will be the unit tests that verify the particular parts of the
config generation that are somewhat complex work properly.