fix(tests): Use more specific comparators in tests
Protobufs support a proto.Equal method that uses proto-supported
reflection to deeply compare two messages; use this instead of
reflect.DeepEqual as go reflection should not be used on protobufs
in favor of protobuf-specific reflection.

Similarly, in one case we're comparing bytes with reflect.DeepEqual
whereas it would be better to just use the specific comparator for
bytes. (I remember thinking this when I wrote it but I guess I
never changed it.)