feat(pkg): include optional service enabled as config property (#67)
We have (for now) decided to have whether or not an optional service is enabled be a property of any services that need that information, rather than service discovery information in spinnaker.yml. Let's translate canary.enabled into services.kayenta.enabled for the two services that need to communicate with Kayenta (Gate and Orca).