refactor(proto): Add baseUrl field to ServiceEnabled and rename
We will in some cases need to have kleat set the baseUrl field
for a service (in cases where the hal config affects this, such
as gate and deck). As this lives under the same config key as
whether the service is enabled, let's augment that proto to
supply both these fields. Also rename it as it's more general
than when it was initially designed.