feat(rosco): Generate rosco config (#69)
* feat(rosco): Add rosco config protos

The rosco config is very similar to the clouddriver config, in
that it contains the configured cloud providers at the top level.
The one difference is that only some cloud providers support
baking, so it only contains these providers.

* feat(rosco): Generate rosco config

This commit adds support for generating a rosco config from a
hal config.

* feat(rosco): Add rosco output to service configs

Also add rosco.yml to the end-to-end test.

* docs(rosco): Add note to changelog about baseImages