fix(deck): Use override URL for gate if specified (#79)
* fix(deck): Use override URL for gate if specified

If an override URL for gate is specified, it should be used in
the deck config.

Also, I updated the default URL to be localhost; as this is used
to set a client-side URL it will be accessed from outside the
cluster where the cluster-local address likely won't be routable.
Setting this to localhost allows you to port forward to the
cluster to access gate.

I'd imagine that anyone actually running a real instance has the
override set (rather than forwarding) but I think localhost is
probably a better default (and is what Halyard does).

* test(deck): Fix and add tests

The last commit broke some tests; fix them, and add a test specifically
testing that the overrideUrl for gate is properly included in the deck
config.

* fix(deck): Remove writing deck.yml

I added this in the last commit to debug a test, and forgot to
remove it.