refactor(core): Move config generation to core
Until we find a valid use case for clients needing to individually
generate service configs, let's move the code to generate the individual
configs to the internal/ folder, and only expose the top-level functions
for generating all configs.

Let's also shorten the name of all of the hal_to_* files.