refactor(core): Add godoc and fix lint errors
This commit adds godoc to all packages in internal/. As part of
this it adds a newline before the copyright and the package name
as I learned that otherwise the copyright statement is interpreted
as a package-level comment.

Most of the lint errors in this directory were just missing comments,
but fixed a couple of other things (in particular capitalizing URL).

Also rename a couple of packages to be less verbose.