style(core): Fix inspections
I ran the IntelliJ inspections, and it found three issues, which
I fixed since I'm on a roll with minor cleanup. They are:
* Explicitly assign the result of fmt.Printf to _ to indicate that
we are ignoring the error (as there's not much point handling it
right before os.Exit anyway). Also use Printf instead of Fprintf
as we aren't formatting
* Rename 'diff' variable which collides with package name
* Change unused 'h' to _ in function signature (though obviously this
will need to be updated when the TODO is handled).