test(core): Improve debuggability of unit tests (#85)
* test(core): Create named type for test cases

All of the config generation tests define the same struct of test
cases inline; let's just give it a name and re-use it across the
tests to reduce duplication.

Also remove halTo... from the variables holding the test cases for
brevity.

* test(core): Add configTest type

This type wraps a config generator and a slice of tests that
test that generator.

Update each of the tests to specify the generator and the actual
tests. As this changes the indentation of all the tests, this
commit will look much cleaner ignoring whitespace.

* test(core): Pull common test logic to a function

All the config generator tests do the exact same thing to a
configTest stuct; pull this logic into a helper function so
it's not repeated in every test file.

* test(core): Improve debuggability of unit tests

If we find that two config protos are not equal, print a diff of
them instead of the full contents of both protos.